+++ CHANGE_LOG.md
@@ -1,7 +1,21 @@
 # Change Log
 
+- Alpha releases (working version, but API still evolving)
+
+## 0.5.2
+
+### Message Function
+
+- Rule `message` can also be a function which receives a `context` object with properties for the message
+  - Current value of the property
+  - Rule name
+  - Property name
+  - Min (for number)
+  - Max (for number)
+  - Min Length (for string)
+  - Max Length (for string)
+
 ## 0.5.1
 
-- Alpha release (working version, but API still evolving)
 - Fully `reactive` Validator object created around the `model` supplied.
-- Working prototype
+- First working prototype of the module.